requirements.txt
anthropic
numpy
pandas
python-dotenv
scikit-learn
sentence-transformers